+++ CHANGELOG.markdown
@@ -1,5 +1,5 @@
-0.3.1 [???]
------
+0.3.1 [2021-01-04]
+------------------
 * Downgraded testing claims that NaNs will round-trip, as 32-bit GHCs aren't fulfilling that promise.
   Now we merely claim that a NaN will return as a NaN.
 * Always provide `NFData Half` instance